% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-registry.R
\name{.fa_make_interface_registry}
\alias{.fa_make_interface_registry}
\title{Construct a complete pairwise interface registry}
\usage{
.fa_make_interface_registry(modules)
}
\arguments{
\item{modules}{Character vector of unique module names.}
}
\value{
A data frame containing all pairwise module interfaces.
}
\description{
Construct a complete pairwise interface registry
}
\keyword{internal}
