% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-definitions.R
\name{.fa_materialize_definition}
\alias{.fa_materialize_definition}
\title{Materialize a stored definition}
\usage{
.fa_materialize_definition(x)
}
\arguments{
\item{x}{A definition list or constructor function.}
}
\value{
A definition list.
}
\description{
Definition registries may contain either a definition list or a
constructor function that returns a fresh definition list.
}
\keyword{internal}
