% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory-overlay.R
\name{.fa_validate_regulatory_overlay}
\alias{.fa_validate_regulatory_overlay}
\title{Validate a regulatory overlay}
\usage{
.fa_validate_regulatory_overlay(regulatory_overlay)
}
\arguments{
\item{regulatory_overlay}{Either \code{NULL} or a named list containing
\code{nodes} and \code{edges} data frames.}
}
\value{
\code{NULL} or a validated regulatory-overlay list.
}
\description{
Validate a regulatory overlay
}
\keyword{internal}
