% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture-definitions.R
\name{.resolve_architecture_definition}
\alias{.resolve_architecture_definition}
\title{Resolve an architecture definition}
\usage{
.resolve_architecture_definition(x = NULL)
}
\arguments{
\item{x}{Either \code{NULL}, the name of a built-in architecture definition,
or a custom architecture-definition list.}
}
\value{
A validated architecture-definition list.

A Broad6 architecture-definition list.
}
\description{
Resolve an architecture definition
}
\keyword{3A}
\keyword{3B}
\keyword{Broad6}
\keyword{Collapses}
\keyword{Level}
\keyword{PHMIES}
\keyword{and}
\keyword{architecture}
\keyword{broad}
\keyword{conserved}
\keyword{core}
\keyword{definition}
\keyword{for}
\keyword{fruit-architecture}
\keyword{functional}
\keyword{internal}
\keyword{into}
\keyword{modules}
\keyword{pathways}
\keyword{reconstruction.}
\keyword{retaining}
\keyword{six}
\keyword{strict}
\keyword{the}
\keyword{while}
