% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportArchitecture}
\alias{exportArchitecture}
\title{Export architecture tables and metadata}
\usage{
exportArchitecture(x, directory, prefix = "fruit_architecture")
}
\arguments{
\item{x}{A \code{fruit_architecture} object.}

\item{directory}{Output directory.}

\item{prefix}{Optional file prefix.}
}
\value{
Invisibly returns the normalized output directory.
}
\description{
Export architecture tables and metadata
}
