% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{fruitArchitecture-package}
\alias{fruitArchitecture-package}
\title{fruitArchitecture: Higher-order signaling architecture reconstruction}
\description{
The package reconstructs annotation-derived modules, interfaces, Level 3A
architectures, and strict Level 3B overlaps from transcriptomic evidence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ensignanalytics/fruitArchitecture}
  \item \url{https://ensignanalytics.github.io/fruitArchitecture/}
  \item Report bugs at \url{https://github.com/ensignanalytics/fruitArchitecture/issues}
}

}
\author{
\strong{Maintainer}: Franklin T. Johnson \email{franklin@ensignanalytics.com}

}
\keyword{internal}
