% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species-registry.R
\name{supportedSpecies}
\alias{supportedSpecies}
\title{List species planned or available in the annotation registry}
\usage{
supportedSpecies()
}
\value{
A data frame describing current registry status.
}
\description{
List species planned or available in the annotation registry
}
