% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture-definitions.R
\name{.default_architecture_definition}
\alias{.default_architecture_definition}
\title{Default PHMIES architecture definition}
\usage{
.default_architecture_definition()
}
\value{
A named list containing module names, aliases, pairwise interfaces,
Level 3A requirements, and metric weights.
}
\description{
Default PHMIES architecture definition
}
\keyword{internal}
