% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture-definitions.R
\name{.resolve_architecture_definition}
\alias{.resolve_architecture_definition}
\title{Resolve an architecture definition}
\usage{
.resolve_architecture_definition(x)
}
\arguments{
\item{x}{Either \code{NULL}, the name of a built-in architecture definition,
or a custom architecture-definition list.}
}
\value{
A validated architecture-definition list.
}
\description{
Resolve an architecture definition
}
\keyword{internal}
